+++ opentelemetry-otlp/src/lib.rs
@@ -687,8 +687,8 @@ mod retry;
 #[cfg(any(feature = "grpc-tonic", feature = "http-proto", feature = "http-json"))]
 mod retry_policy;
 
-pub use crate::exporter::Compression;
 pub use crate::exporter::ExporterBuildError;
+pub use crate::exporter::{Compression, ParseCompressionError};
 #[cfg(feature = "trace")]
 #[cfg(any(feature = "http-proto", feature = "http-json", feature = "grpc-tonic"))]
 pub use crate::span::{